Version 2.0.2
* Using full path when running real QEMU binary in a chroot
* Fixed help line in vnetctl "update-conf" command
* Removed debug messages
* Added command 'vmm info' -- details in human-readable format
* Small fix: the name of the field with IP addresses (ips => addrs) in config_network
* Bugfix: incorrect name resolution from the machine's home dir
* Bugfix: in GetRouteTableIndex() function the following files are now scanned:
  - /etc/iproute2/rt_tables
  - /usr/share/iproute2/rt_tables
  - /etc/iproute2/rt_tables.d/*
* Fixed work with the net_cls cgroup v1 controller
* Replace outbound traffic shaping from cgroups net_cls to Traffic Control IP filters
* Use uint64 for memory values
* Added vnetctl info/inspect command
* Bugfix: live migration fails if VM has UEFI+pflash
* Use io_uring as AIO backend for cdrom devices
* Bugfix: machine instance registration collision
* Add new config parameter 'vga'
* Remove rsyslog config for kvmrund service

+++ services/system/service_protoconv.go
@@ -40,7 +40,7 @@ func incomingRequisitesToProto(requisites *system.IncomingRequisites) *pb_types.
 
 func optsFromQemuInstanceRegisterRequest(req *pb.QemuInstanceRegisterRequest) *system.InstanceRegistrationOptions {
 	return &system.InstanceRegistrationOptions{
-		MemActual: uint32(req.MemActual),
+		MemActual: req.MemActual,
 		PID:       uint32(req.PID),
 	}
 }
